Improve /Wall handling
MSVC's /Wall is not great. A lot of standard library & Windows headers
will generate warnings. So it should not be used for regular dev, in my
opinion.

However, it is nice to check in on it from time to time to see if
anything might be unintentionally risky.

This commit addresses all the reasonable issues that /Wall showed.
Some warnings were left as they're basically useless (such as a warning
that padding bytes will be added to the end of a class).

+++ Code/maxGUI/Form.cpp
@@ -178,8 +178,8 @@ namespace maxGUI {
 		DWORD extra_style = WS_EX_CONTROLPARENT | WS_EX_CLIENTEDGE; //WS_EX_WINDOWEDGE;
 
 		AdjustWindowRectEx(&area, win32_styles, FALSE, extra_style);
-		DWORD total_height = 0;
-		DWORD total_width = 0;
+		LONG total_height = 0;
+		LONG total_width = 0;
 		if (virtual_top == CW_USEDEFAULT)
 		{
 			area.bottom -= area.top;
